Value Inc Sales.py
The Sales Manager has no sales reporting but he has a brief idea of current sales. He also has no idea of the monthly cost, profit, and top-selling products. with the help of Python, I used to clean the data and find the insights and patterns. after analyzing we used Tableau to create the sales dashboard.